requirements.txt
sqlalchemy
flask
flask-sqlalchemy
flask-migrate
pyjwt
requests
flask-cors